fix: standardize event names across classes, methods, and events
