+++ messages/packageVersionCreate.md
@@ -73,3 +73,7 @@ post-install URL
 # malformedUrl
 
 The %s value "%s" from the command line or sfdx-project.json is not in the correct format for a URL. It must be a valid URL in the format "http://salesforce.com". More information: https://nodejs.org/api/url.html#url_url_strings_and_url_objects
+
+# versionCreateFailedWithMultipleErrors
+
+Multiple errors occurred:
