+++ src/package/packageInstall.ts
@@ -9,15 +9,14 @@ import { Connection, Lifecycle, Logger, Messages, PollingClient, SfError, Status
 import { isString, isNumber, Optional } from '@salesforce/ts-types';
 import { QueryResult } from 'jsforce';
 import { Duration } from '@salesforce/kit';
-import { PackagingSObjects } from '../interfaces';
 import {
   isErrorFromSPVQueryRestriction,
   isErrorPackageNotAvailable,
   getPackageTypeBy04t,
   escapeInstallationKey,
-} from '../utils/packageUtils';
+} from '../utils';
 import { consts } from '../constants';
-import { PackageInstallOptions, PackageInstallCreateRequest } from '../interfaces/packagingInterfacesAndType';
+import { PackagingSObjects, PackageInstallOptions, PackageInstallCreateRequest } from '../interfaces';
 
 import SubscriberPackageVersion = PackagingSObjects.SubscriberPackageVersion;
 import PackageInstallRequest = PackagingSObjects.PackageInstallRequest;
@@ -58,21 +57,21 @@ export async function installPackage(
   if (pkgType !== 'Unlocked') {
     if (request.UpgradeType !== defaults.UpgradeType) {
       const msg = installMsgs.getMessage('upgradeTypeOnlyForUnlockedWarning');
-      await Lifecycle.getInstance().emit('PackageInstallRequest:warning', msg);
+      await Lifecycle.getInstance().emit('Package/install-warning', msg);
       delete request.UpgradeType;
     }
     if (request.ApexCompileType !== defaults.ApexCompileType) {
       const msg = installMsgs.getMessage('apexCompileOnlyForUnlockedWarning');
-      await Lifecycle.getInstance().emit('PackageInstallRequest:warning', msg);
+      await Lifecycle.getInstance().emit('Package/install-warning', msg);
       delete request.ApexCompileType;
     }
   }
 
-  await Lifecycle.getInstance().emit('PackageInstallRequest:presend', request);
+  await Lifecycle.getInstance().emit('Package/install-presend', request);
 
   const result = await connection.tooling.create('PackageInstallRequest', request);
 
-  await Lifecycle.getInstance().emit('PackageInstallRequest:postsend', result);
+  await Lifecycle.getInstance().emit('Package/install-postsend', result);
 
   const packageInstallRequestId = result.id;
   if (!packageInstallRequestId) {
@@ -165,7 +164,7 @@ async function pollStatus(
     poll: async (): Promise<StatusResult> => {
       packageInstallRequest = await getStatus(connection, installRequestId);
       getLogger().debug(installMsgs.getMessage('packageInstallPolling', [packageInstallRequest?.Status]));
-      await Lifecycle.getInstance().emit('PackageInstallRequest:status', packageInstallRequest);
+      await Lifecycle.getInstance().emit('Package/install-status', packageInstallRequest);
       if (['SUCCESS', 'ERROR'].includes(packageInstallRequest?.Status)) {
         return { completed: true, payload: packageInstallRequest };
       }
@@ -229,14 +228,14 @@ export async function waitForPublish(
       let installValidationStatus: SubscriberPackageVersion['InstallValidationStatus'];
       if (queryResult?.records?.length) {
         installValidationStatus = queryResult.records[0].InstallValidationStatus;
-        await Lifecycle.getInstance().emit('SubscriberPackageVersion:status', installValidationStatus);
+        await Lifecycle.getInstance().emit('Package/install-status', installValidationStatus);
         if (!['PACKAGE_UNAVAILABLE', 'UNINSTALL_IN_PROGRESS'].includes(installValidationStatus)) {
           return { completed: true, payload: queryResult };
         }
       }
       const tokens = installValidationStatus ? [` Status = ${installValidationStatus}`] : [];
       getLogger().debug(installMsgs.getMessage('publishWaitProgress', tokens));
-      await Lifecycle.getInstance().emit('SubscriberPackageVersion:status', installValidationStatus);
+      await Lifecycle.getInstance().emit('Package/install-status', installValidationStatus);
       return { completed: false, payload: queryResult };
     },
   };
